requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
boilerpy3==1.0.7
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
cryptography==45.0.5
distro==1.9.0
docopt==0.6.2
Events==0.5
exceptiongroup==1.3.0
farm-haystack==1.26.4.post0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.23.2
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy-imports==0.3.1
lit==18.1.8
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.3.1
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.3.8
posthog==6.1.0
prompthub-py==4.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-cache==0.9.8
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.15.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.47.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.20.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.46.0
triton==2.0.0
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
weaviate-client==3.26.7
websockets==15.0.1
